support max_by/min_by for spark
fix format

skip fuzzer

stage

add doc

fix build

fix comments

fix format

fix format
