+++ velox/functions/sparksql/aggregates/CMakeLists.txt
@@ -14,7 +14,7 @@
 add_library(
   velox_functions_spark_aggregates
   BitwiseXorAggregate.cpp BloomFilterAggAggregate.cpp FirstLastAggregate.cpp
-  AverageAggregate.cpp Register.cpp)
+  AverageAggregate.cpp MinMaxByAggregate.cpp Register.cpp)
 
 target_link_libraries(velox_functions_spark_aggregates fmt::fmt velox_exec
                       velox_expression_functions velox_aggregates velox_vector)
